requirements.txt
phidata 
sqlalchemy
psycopg[binary]
pgvector
fastapi[standard]
pypdf